Extracting @groups to #groups method
Using instance variables in views creates strong coupling with
corresponding controllers. Ideally all instance variables in the view
would be replaced with corresponding helper methods.

In doing so, we would be dramatically improving our ability to modify
views and controllers.

At present we are looking to modify the batch_edits controller to prompt
users for different forms based on a selected object type (i.e. law
journal or senior thesis or ETD). There is a larger refactoring that
would need to occur, but this is my initial attempt to fix one of the
issues that we will be having.