Checksum_audit_logs updated to account for changes in Fedora 4.
Notice that ChecksumAuditLog.prune_history now expects a pid rather than a version as the parameter.

In addition, Audit.logs() now does not expect a parameter.

Conceptually, the bigger change is that now ChecksumAuditLog will NOT include the datastream ID when it saves the data. In the pass, the datastream IDs was a parameter (although it was always 'content') and now we are not passing this parameter. The MySQL table checksum_audit_logs remains unchanged. We could later on update this code to handle multiple datastreams if we want to.