Remove logic around user groups from ability methods.
The current code forces downstream adopters to override entire ability methods, such as #editor_abilities, if they use a different strategy for determining if a user is an admin/registered user. Case in point: ScholarSphere overrides #editor_abilities, and because of this, when that method needs to change in Sufia, ScholarSphere's method needs to be updated manually to pick up the change. Instead, abstract away *how* users are determined to be registered/admin into their OWN methods.