Make my controller retrieve only user's own collections.
My controller used to get all collections from the system rather
than just the ones belonging to the user. When adding files to
a collection, this would give the user a list of all collections
to choose from. The user couldn't add the files to someone else's
collection anyway so those options were useless.