After a week of training on "Practical Object-Oriented Design" with Sandi Metz @cam156 @awead and I decided to take a stab at refactoring a small piece of functionality with some of the ideas that we learned.
This PR does not change the functionality of the MailboxController but it refactors most of the code out of it to a model class UserMailbox This new class encapsulates most of the functionality and the controller is now as skinny as it gets which was one of our goals.

The new UserMailbox has a more comprehensive list of tests than the original code and it's easier to test since it does not mix model + controller functionality.