Use sufia.root_path instead of root_path
In the navigation bar, the `Home` link and logo go to `root_path`.
When Sufia is mounted as an engine at a sub-path, this will take the
user to the root_path of the host Rails application rather than back to
the main Sufia page.

This change makes it so that the Home link and logo navigate to Sufia’s
root_path instead, so that the Sufia interface is more self-contained.