Hydra-Jetty should be the default service container in 6.x.
We could alternatively allow AF to >= 9.8.0 but then we'll need to make changes to the code and documentation, or we could switch 6.x to using fcrepo_wrapper and solr_wrapper by default, but those will be new to 6.x developers. Therefore, for now, to enable new and existing Sufia development to just work with the same hydra-jetty assumptions as before, just pin AF to < 9.8.0 and the problems and extra work go away.